+++ reme/components/service/__init__.py
@@ -1,15 +1,13 @@
 """Service components for exposing jobs via different protocols."""
 
 from .base_service import BaseService
-from .cli_service import CliService, prepare_start_config, should_precheck_start
+from .cli_service import CliService
 from .http_service import HttpService
 from .mcp_service import MCPService
 
 __all__ = [
     "BaseService",
     "CliService",
-    "prepare_start_config",
-    "should_precheck_start",
     "HttpService",
     "MCPService",
 ]
